requirements.txt
argcomplete==3.1.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
autocorrect==2.6.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama @ file:///C:/ci_311/colorama_1676422310965/work
coloredlogs==15.0.1
comm @ file:///C:/ci_311/comm_1678376562840/work
contourpy==1.2.0
cramjam==2.7.0
cycler==0.12.1
dask==2023.11.0
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.7.4
fast-sentence-transformers==0.4.1
fastparquet==2023.10.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
fsspec==2023.10.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
ipykernel @ file:///C:/b/abs_07rkft_vaz/croot/ipykernel_1691121700587/work
ipython @ file:///C:/b/abs_e5729i179y/croot/ipython_1694181400005/work
jedi @ file:///C:/ci_311/jedi_1679427407646/work
Jinja2==3.1.2
joblib==1.3.2
jupyter_client @ file:///C:/b/abs_a6h3c8hfdq/croot/jupyter_client_1699455939372/work
jupyter_core @ file:///C:/b/abs_c769pbqg9b/croot/jupyter_core_1698937367513/work
kiwisolver==1.4.5
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
mpmath==1.3.0
nest-asyncio @ file:///C:/ci_311/nest-asyncio_1676423519896/work
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
onnx==1.15.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==2.1.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.4.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.1.0
pipx @ git+https://github.com/pypa/pipx.git@f5bff4276572f6e5c3b143a06d9a826dc86bcaea
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
prompt-toolkit @ file:///C:/ci_311/prompt-toolkit_1676425940920/work
protobuf==4.25.1
psutil==5.9.6
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==14.0.1
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3.post1
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
threadpoolctl==3.2.0
tk==0.1.0
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado @ file:///C:/b/abs_0cbrstidzg/croot/tornado_1696937003724/work
tqdm==4.66.1
traitlets @ file:///C:/ci_311/traitlets_1676423290727/work
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
userpath==1.9.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
zipp==3.17.0
